refactor: 74 - move links into functions
