+++ src/components/registration/layout.ts
@@ -1,12 +1,13 @@
 import { REGISTRATION } from '../../pages/registration/constants';
-import { registrationLink } from '../../shared/components/link';
+import { LINK_CONFIG } from '../../shared/components/link';
 import { HEADER2 } from '../../shared/styles';
-import { createDiv, createH2 } from '../../utils/create-elements/create-tags';
+import { createA, createDiv, createH2 } from '../../utils/create-elements/create-tags';
 import { form } from './form';
 
 export function registrationLayout(): HTMLDivElement {
   const title = createH2({ classes: HEADER2.general, text: 'Registration' });
   const registrationForm = form();
+  const registrationLink = createA(LINK_CONFIG.registration);
 
   const layout = createDiv({
     children: [title, registrationForm, registrationLink],
